advection-based redistribution of molecules in CPM
When the domain changes during the CPM_step, we now move molecules
around according to the advection equation.

A radial velocity field, where |v(r)|=lambda/r^2, pointing
inward/outward of the expanding/contracting point is assumed, and
molecules are resdistributed in a probabilistic manner that is in
accordance with the advection equation.